refactor task and attachment controllers with improved commands and queries

+++ Application/Features/Tasks/Commands/Delete/DeleteTaskCommandHandler.cs
@@ -2,6 +2,8 @@ using TaskFlow.Application.Common;
 using TaskFlow.Application.DTOs.Task;
 using TaskFlow.Application.Interfaces.Messaging;
 using TaskFlow.Application.Interfaces.Repositories;
+using TaskFlow.Application.Interfaces.Services;
+using TaskFlow.Domain.Enums;
 using TaskFlow.Domain.Services;
 
 namespace TaskFlow.Application.Features.Tasks.Commands;
@@ -11,15 +13,18 @@ public class DeleteTaskCommandHandler
 {
     private readonly IProjectRepository _projectRepository;
     private readonly ITaskRepository _taskRepository;
+    private readonly IFileService _fileService;
     private readonly IProjectPermissionService _permissionService;
     public DeleteTaskCommandHandler(
         IProjectRepository projectRepository,
         ITaskRepository taskRepository,
+        IFileService fileService,
         IProjectPermissionService permissionService
     )
     {
         _projectRepository = projectRepository;
         _taskRepository = taskRepository;
+        _fileService = fileService;
         _permissionService = permissionService;
     }
     public async Task<Result<TaskDeleteDto>> Handle(
@@ -36,6 +41,14 @@ public class DeleteTaskCommandHandler
 
         if (!_permissionService.CanDeleteTask(project, request.InitiatorId, task))
             return Result<TaskDeleteDto>.Failure(new Error("User.InsufficientPermissions", $"User cannot delete this task"));
+
+        var deleteTaskDirectoryResult = await _fileService.DeleteEntityDirectoryAsync(
+            ContentType.Tasks,
+            task.Id,
+            cancellationToken);
+
+        if (!deleteTaskDirectoryResult)
+            return Result<TaskDeleteDto>.Failure(new Error("FileService.FailedToDelete", $"Failed to delete task '{task.Id}' content directory"));
 
         await _taskRepository.RemoveTaskAsync(task, cancellationToken);
         await _taskRepository.SaveChangesAsync(cancellationToken);
